pyproject.toml
[project]
name = "estudo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.6",
    "fastapi>=0.116.1",
    "groq>=0.30.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
